CT: Perf optimization through prefetch and condition checks in CT
JIRA: SAMPLEVNF-10

LLC miss is possible when the connection data structure is accessed
after retrieved from hash table entries for TCP/UDP.
CPI for rte_ct_udp_packet() is not good due to if/else condition checks.
This patch provide a prefetch instructions to pre-fetch the cnxn data entry
structure to avoid the LLC miss and optimize the if/else condition check
to improve the CPI for rte_ct_udp_packet() function.

Change-Id: Ieaf70815c919ddbe1ff34921222eac030fb41756
Signed-off-by: Anand B Jyoti <anand.b.jyoti@intel.com>