vCGNAPT: correcting to use default rte_ring_dequeue
JIRA: SAMPLEVNF-19

The rte_ring_sc_dequeue returns ENOENT in DPDK16.04 while ENOBUFS in
DPDK17.05. This leads to error in return value checking and mis behaviour.

Using of rte_ring_sc_dequeue() to be avoided to use the default
configuration as set during the creation of the queue as per the
DPDK API documentation.

Change-Id: Iacee1349b26d6ab432be891fad12313a6d68ca4d
Signed-off-by: Anand B Jyoti <anand.b.jyoti@intel.com>