arp: Config parameters longer the PIPELINE_MAX_ARGS chars ignored
JIRA: SAMPLEVNF-X

Due to this KW fix the arp pipeline is rejecting the commands
greater than PIPELINE_MAX_ARGS size. This is seen in arp_route_tbl
which can go long in case of multi port setup.

This patch undo the KW fix which is causing issue with longer
parameters under ARPICMP pipeline.

Change-Id: I96ba3882b5fecc6ac2f7b75f11a63083b847bba5
Signed-off-by: Anand B Jyoti <anand.b.jyoti@intel.com>