arp: Config parameters longer than PIPELINE_MAX_ARGS chars ignored
JIRA: SAMPLEVNF-86

Due to this KW fix the arp pipeline is rejecting the commands
greater than PIPELINE_MAX_ARGS size. This is seen in arp_route_tbl
which can is becoming bigger in case of multi port setup.

This patch undo the KW fix which is causing issue with longer
parameters under ARPICMP pipeline.

Change-Id: I7bd7124118f531aa70d0fddd75a8ae693455d361
Signed-off-by: Anand B Jyoti <anand.b.jyoti@intel.com>