lw_AFTR: IP Checksum required on generated packet.
In lwAFTR (ipv6 de-encapsulate) the generated ipv4 packet requires an IP
checksum, otherwise this packet is dropped in Openstack. (Openstack is
much stricter than baremetal) Some fields in the header need to be
initialized.

Add flag in Makefile which when enabled will build to include the generate
the IP checksum in the build. FLAG = GEN_DECAP_IPV6_TO_IPV4_CKSUM=y
Not include in default build.

Change-Id: If94e7cff64b03c66362021f05e48fb9265fc8210
Signed-off-by: Daniel Martin Buckley <daniel.m.buckley@intel.com>