Add config option to use port mac as src mac in l2fwd and swap
L2fwd and swap has many options to set src mac. swap was supposed
to support the ability to use port mac, but config file parsing was
wrong. L2fwd was using port mac if a port was available and if
src mac=packet or if there was no src mac in the config.

Now l2fwd supports the following options
- "src mac=xx:xx:xx:xx:xx:xx" => this mac address is used as src mac.
- "src mac=packet" => the src mac is taken from the dst mac of the
	received packet.
- "src mac=hw" => the src mac is taken from the mac address of the port,
	if there is a physical port. Error otherwise.
- "src mac=no" => src mac kept untouched
- No "src mac" => same as "src mac=hw" if there is a physical port
              and same as "src mac=packet" otherwise.

Default is (no "src mac") hence the mac is taken from the tx port if there
              is one tx port and from the packet otherwise.

swap support is similar, except that it does not support "src mac=no".

Change-Id: I70fe49a61c2e85772288b769ede14a7a6205d122
Signed-off-by: Xavier Simonart <xavier.simonart@intel.com>