Update patch set 1
Patch Set 1: Code-Review+1

Seems to work for the current problem at hand, so voting +1 as is.
Ideally, this should be handled inside parse_list_set(), in order to also work with DATS lwAFTR traffic generation and with "tx cores=" lists used in complex packet processing workloads.
Not sure it's possible to define $var1=# in the [variables] or [lua] sections, although it works with the -w command-line option. So I would recommend that we advertise this feature as: $var1=none. With this syntax in place, we can later rework parse_list_set() to deal with other use cases.

Patch-set: 1
Reviewer: Gerrit User 7067 <7067@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+1