Add check for configuration error (not enough mbuf)
Number of mbufs must at least cover for number of rx descriptors,
number of tx descriptors, number of mbuf cached and number of
mbufs handled by the application.

If this is not the case, a warning is returned.

This ony check for the more basic cases. This will not check for
instance for cases with multiples rings where more mbufs might be
cached.

Change-Id: If2c0c9fc76ed4500849d92cf7586bb0b25d8ab22
Signed-off-by: Xavier Simonart <xavier.simonart@intel.com>