Support multiple tasks in "dp core stats" command
The "dp core stats" command supported, as other commands, multiple cores
but not multiple tasks. It now supports both (e.g. dp core stats 1-2 0-1).
This is necessary so that users can collect all statistics from PROX in one
command line.

It will return similar information on the screen or through the socket:
- if the syntax is wrong (e.g. missing task, too many cores or tasks, ...)
  => an error message is printed on the screen, and the single error line
     "error: invalid syntax" is returned through the socket
- if the syntax is correct
  => one line is printed or returned for each core/task pair
     => if the core/task pair is invalid, the line reports an error
     => otherwise, the line provides dataplane statistics for the core/task pair,
        with the core_id and task_id added, either at the end of the line returned
        through the socket, or at the beginning of the line printed on the screen.

This means a change in behaviour when using the socket: before, when an error
happened (e.g. bad task), then nothing was returned which made it difficult for
the script to catch the syntax error.

Similar behaviour could/should be implemented for other commands.
However, care must be taken as some scripts like NSB might count the number
of lines within the output, or the number of items within the line.
This should not be an issue for dp core stats as NSB does not use it, and other
PROX scripts do not count the number of items and will be adjusted to count the
number of lines.

Change-Id: I582a671ae3d2f6493f791e80fc28e70f8e3a38d1
Signed-off-by: Xavier Simonart <xavier.simonart@intel.com>
Signed-off-by: Patrice Buriez <patrice.buriez@intel.com>