Support multiple tasks in dp core stats command
dp core stats command supported, as the other commands, multiple cores
but not multiple tasks. it now wupports both (e.g. dp core stats 1-2 0-1).
This is necessary so that users can collect all statistics from PROX in one
command line.

It will return similar information on the screen or through the socket:
- if the syntax is wrong (e.g. missing task, too many cores or tasks, ...)
  => one error will be printed or returned through the socket (Invalid syntax)
- if the syntax is correct
  => one line will be printed or returned by core/task pair
     => if the core/task pair is invalid, then an error is returned
     => if the core/task pair is valid, then the statistics are returned
This means a change in behaviour when using the socket: before, when an error
happened (e.g. bad task), then nothing was returned which made it difficult for
the script to catch the bad bahaviour.

In addition, core_id and task_id are added at the end of the statistics line.

Similar behaviour could/should be implemented for other commands.
However, care must be taken as some scripts like NSB might count the number
of lines within the output, or the number of items within in line.
This should not be an issue for dp core stats as NSB does not use it, and other
PROX scripts do not count number of outputs.

Change-Id: I582a671ae3d2f6493f791e80fc28e70f8e3a38d1
Signed-off-by: Xavier Simonart <xavier.simonart@intel.com>