Remove UUIDs in centos.json
Removed the remaining UUIDs from the centos.json file, making it easier
to reuse this file. This does require a recent version of the packer
tool. In the README, also added the --env option in the packer alias, so
that the OS_CLOUD variable is set when running packer in the container.
Updated rapid.vms with the security group name used in the centos.json
file.

Change-Id: I4f6815a3250811e336c14857e0ec5e7793a77acf
Signed-off-by: Luc Provoost <luc.provoost@intel.com>