Allow latency thresholds for TST009 testing
Even if the latency thresholds were defined in the TST009 test files,
the values were ignored and replaced by 'inf', whcih means the
thresholds were not taken into account to define a successful step. If
the thresholds are not defined in the test files, they are still set to
'inf'.

Change-Id: I9559f05fb6066ab8759900abe3eb53ff4662e510
Signed-off-by: Luc Provoost <luc.provoost@intel.com>