deploycentostools.sh to use latest PROX
Rapid scripts are communicating to the PROX instances. Rapid scripts
will not work with every PROX version sice sometime we change the
interface between the scripts and PROX. Hence we have a commit ID in
deploycentostools.sh to make sure we have a version of PROX that we can
work with. This change will now ensure we have a recent PROX version
when we use deploycentostools.sh to build an image.

Change-Id: Ia634f61a2dc8f49188cec6697ebf5f980d35ce7d
Signed-off-by: Luc Provoost <luc.provoost@intel.com>