New PROX version, background traffic reporting
deploycentostools.sh is now installing a recent PROX version. When using
background traffic, the tool is also reporting the total traffic that
was handled by the NFVi stack, including foreground & background
traffic.
The use of the prox_socket and prox_launch_exit was also broken in a
previous release and is not fixed.
When setting prox_socket to false for a PROX instance, we will not
create a socket connection to control that PROX instance. This is not
meaningful for the generator PROX instances since we cannot control the
behavior of PROX during the test. You could use however the generator
against another non-PROX instance. Default is true.
prox_launch_exit is used to start and stop the PROX pogram inside the
instance. Default is true. This is useful in case you want to start PROX
manually to be able to inspect the PROX UI, but at the same time, let
the rapid scripts control the testing.

Change-Id: Ib5aa809f4be201859542769f5f55f4989dad97ef
Signed-off-by: Luc Provoost <luc.provoost@intel.com>