Fix the default test file
In a previous change, all test files were moved to the tests
subdirectory. When not specifying the test file on the command line as a
parameter, the default test file is used, as specified in
rapid_defaults.py
It did not reflect the fact that all test files were moved to
this directory.

Change-Id: I5c77b75dbdde040d0e30dda2b97e981dff572f2f
Signed-off-by: Luc Provoost <luc.provoost@intel.com>