Remapping table for ingress classifier
Added a remapping table for ingress classifier (specified using
flex_offset and flex_width). The size of the table is configurable
with compile-time option CLASSIF_REMAPPING_TABLE_SIZE and disabled
by default. When enabled, the table entries are initialized with the
value of the entry itself i.e. the net effect is no remapping.
The purpose is merely the memory access effect of a remapping table.

Change-Id: Ia063b39af1af4b404209ec9bc88c35ef65faf099
Signed-off-by: Luc Provoost <luc.provoost@intel.com>