Allow to collect power
In order to collect pwoer of the system under test, one can now define a
power helper function that will retrieve System, CPU, and RAM power. This
function will then be called at the end of a test iteration. The actual
function to collect the power is not provided in this github.

Signed-off-by: Luc Provoost <luc.provoost@gmail.com>
Change-Id: Ic33881326f9318f1669fdef69a872596cc8e6dbc